Updated method signatures Added new classes for handling Addmusic Checkbits Moved Addmusic Logic from RomOperations to AddmusicLogic Added CleanRom and supporting functions Added new supporting data
