+++ Addmusic2/Model/Interfaces/IRomOperations.cs
@@ -11,6 +11,8 @@ namespace Addmusic2.Model.Interfaces
 
         int FindFreeSpaceInROM(Rom rom, int size, int start);
 
+        void ClearRATSTag(ref Rom rom, int offset);
+
         int SNESToPC(int address);
         int SNESToPC(int address, bool useSA1);
         int PCToSNES(int address);
